+++ 1.19/Common/src/main/java/fuzs/thinair/advancements/AirProtectionSource.java
@@ -1,0 +1,12 @@
+package fuzs.thinair.advancements;
+
+/**
+ * How were you protected from the air?
+ */
+public enum AirProtectionSource {
+    NONE,
+    INHERENT,
+    BLADDER,
+    RESPIRATOR,
+    WATER_BREATHING,
+}