+++ 1.19/Common/src/main/java/fuzs/thinair/capability/AirProtectionCapability.java
@@ -1,0 +1,10 @@
+package fuzs.thinair.capability;
+
+import fuzs.puzzleslib.capability.data.CapabilityComponent;
+
+public interface AirProtectionCapability extends CapabilityComponent {
+
+    boolean isProtected();
+
+    void setProtected(boolean usingBladder);
+}