+++ 1.19/CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v4.0.1-1.19.2] - 2023-07-08
+### Added
+- Added support for Curios and Trinkets mods
+### Fixed
+- Fixed respirator not working when equipped
+
 ## [v4.0.0-1.19.2] - 2023-06-25
 - Ported to Minecraft 1.19.2